# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API EduApp",
        "description": "Documentation Swagger de l'API d'authentification",
        "version": "1.0.0"
    },
    "paths": {
        "/api/enseignants": {
            "get": {
                "tags": [
                    "Absences"
                ],
                "summary": "Récupérer la liste des enseignants",
                "description": "Retourne tous les enseignants avec leur id et nom",
                "operationId": "21a25eed477bd960dc31e699334ec5cf",
                "responses": {
                    "200": {
                        "description": "Liste des enseignants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "John Doe"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/absences/enseignant/{id}": {
            "get": {
                "tags": [
                    "Absences"
                ],
                "summary": "Récupérer les absences d'un enseignant",
                "description": "Retourne les absences (séances non réalisées) d'un enseignant",
                "operationId": "55ca39c9a5cb311ec38316031e773780",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'enseignant",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des absences",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "enseignant_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_absences": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "dates_absences": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date_seance": {
                                                        "type": "string",
                                                        "example": "2026-01-10"
                                                    },
                                                    "heure_seance": {
                                                        "type": "string",
                                                        "example": "08:00-10:00"
                                                    },
                                                    "nature": {
                                                        "type": "string",
                                                        "example": "Cours"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/absence": {
            "post": {
                "tags": [
                    "Absences"
                ],
                "summary": "Marquer une absence",
                "description": "Créer ou mettre à jour une absence pour un étudiant",
                "operationId": "a41c3318723201002498cb73c2e00fa7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code_etudiant": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "code_matiere": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "code_enseignant": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "seance": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "Absent",
                                            "Present"
                                        ],
                                        "example": "Absent"
                                    },
                                    "justifie": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "date_absence": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-01-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Absence enregistrée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Absence enregistrée avec succès"
                                        },
                                        "absence": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/absences/{id}": {
            "put": {
                "tags": [
                    "Absences"
                ],
                "summary": "Mettre à jour une absence",
                "description": "Modifier le statut ou la justification d'une absence existante",
                "operationId": "49505538b1618c38db992592ef3477c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'absence",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "Absent",
                                            "Present"
                                        ],
                                        "example": "Present"
                                    },
                                    "justifie": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Absence mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Absence mise à jour avec succès"
                                        },
                                        "absence": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Absence introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}